Convert to a custom resource
Also cleanup testing failures
