+++ .travis.yml
@@ -20,17 +20,13 @@ services: docker
 env:
   matrix:
   - INSTANCE=default-centos-5
-  - INSTANCE=default-centos-5 CHEF_VERSION=12.1.2
   - INSTANCE=default-centos-6
-  - INSTANCE=default-centos-6 CHEF_VERSION=12.1.2
+  - INSTANCE=default-centos-6 CHEF_VERSION=12.7.2
   - INSTANCE=default-centos-7
-  - INSTANCE=default-centos-7 CHEF_VERSION=12.1.2
+  - INSTANCE=default-centos-7 CHEF_VERSION=12.7.2
   - INSTANCE=default-ubuntu-1204
-  - INSTANCE=default-ubuntu-1204 CHEF_VERSION=12.1.2
   - INSTANCE=default-ubuntu-1404
-  - INSTANCE=default-ubuntu-1404 CHEF_VERSION=12.1.2
   - INSTANCE=default-ubuntu-1604
-  - INSTANCE=default-ubuntu-1604 CHEF_VERSION=12.1.2
 before_script:
   - sudo iptables -L DOCKER || ( echo "DOCKER iptables chain missing" ; sudo iptables -N DOCKER )
   - eval "$(chef shell-init bash)"
