+++ metadata.rb
@@ -7,7 +7,7 @@ long_description 'Installs/Configures packagecloud.io repositories.'
 version '0.3.0'
 source_url 'https://github.com/computology/packagecloud-cookbook'
 issues_url 'https://github.com/computology/packagecloud-cookbook/issues'
-chef_version '>= 12.1' if respond_to?(:chef_version)
+chef_version '>= 12.5' if respond_to?(:chef_version)
 %w(ubuntu debian redhat centos amazon oracle fedora scientific).each do |p|
   supports p
 end
