+++ spec/matchers_spec.rb
@@ -2,7 +2,7 @@ require 'chefspec'
 require 'chefspec/berkshelf'
 
 describe 'packagecloud_test::deb' do
-  let(:chef_run) { ChefSpec::SoloRunner.converge(described_recipe) }
+  let(:chef_run) { ChefSpec::SoloRunner.new(platform: 'ubuntu', version: '16.04').converge(described_recipe) }
 
   it 'should create packagecloud_repo' do
     expect(chef_run).to create_packagecloud_repo('computology_public_deb')
