+++ test/fixtures/cookbooks/packagecloud_test/recipes/deb.rb
@@ -16,7 +16,6 @@ end
 package 'jake-doc'
 
 execute 'install_jake_source' do
-  cwd '/home/vagrant'
   command 'apt-get source jake'
 end
